Add copying information
For more information on why I use 🅭🄍, see
<https://jasonyundt.website/posts/why-do-i-use-cc0.html>.

While progress has been made [1] into closing the
🅭🄍-not-available-under-🅭🄍 loophole [2] [3], it’s not fully closed.
Creative Commons assigned someone to one of the issues, and that
assignee said that 🅭🄍 __should__ already be available under 🅭🄍 [1].
This, unfortunately, isn’t quite the same as Creative Commons outright
saying that 🅭🄍 __is__ available under 🅭🄍. Additionally, some text was
added to CC’s Licensing Statement for Content and Software Code [4]
which dedicates “the text of all of the notices and accompanying text on
the license pages” to the public domain using 🅭🄍. The key words here are
“on the license pages”. 🅭🄍 is not a license [5] so that quote,
unfortunately, doesn’t apply to 🅭🄍. Ugh.

With that in mind, I decided to still include attribution text for the
portions of 🅭🄍 that are (potentially) 🅭🅯4.0 licensed. It’s probably not
necessary, but I’d rather err on the side of caution.

The content that this commit adds is in plain text. In a future commit,
I might turn some of it into Markdown, but I want to set up pre-commit
first [6].

[1]: <https://github.com/creativecommons/creativecommons.org/issues/1218#issuecomment-1212525987>
[2]: <https://github.com/creativecommons/creativecommons.org/issues/1212>
[3]: <https://github.com/creativecommons/creativecommons.org/issues/1218>
[4]: <https://creativecommons.org/policies#licensing%20statement>
[5]: <https://creativecommons.org/faq/#how-do-cc-licenses-operate>
[6]: <https://pre-commit.com/>